! Improved maintenance: allow moving old topics based on number of days, picking source boards, that kind of thing. I even tried to get the translation right. (Oh, yes, I did also clean up the strings slightly to fix compounding issues.) (ManageMaintenance.php, ManageMaintenance.template.php, ManageMaintenance language file) ! Remove bulk topics: there are two kinds of locked topics, not accounted for in SMF either, also in our case it's possible that the remove topics function could be called, legitimately, for a topic that no longer exists. This should be able to be handled safely without an error (intentionally) (RemoveTopic.php) ! When approving posts, the last message id needs to be recalculated otherwise weird things can happen, like messages you can't actually mark as read. (Subs-Post.php) ! Post moderation status needs to be maintained more than before now there are other ways to get posts moderated that aren't related to the moderation filters. (Load.php, ManageModeration.php, Post2.php)

+++ Sources/RemoveTopic.php
@@ -178,8 +178,8 @@ function RemoveOldTopics2()
 	elseif ($_POST['delete_type'] == 'locked')
 	{
 		$condition .= '
-			AND t.locked = {int:locked}';
-		$condition_params['locked'] = 1;
+			AND t.locked = {int:unlocked}';
+		$condition_params['unlocked'] = 0; // There are two kinds of locked topics.
 	}
 
 	// Exclude pinned?
@@ -497,14 +497,17 @@ function removeTopics($topics, $decreasePostCount = true, $ignoreRecycling = fal
 			$messages[] = $row['id_msg'];
 		wesql::free_result($request);
 	}
-	wesql::query('
-		DELETE FROM {db_prefix}likes
-		WHERE id_content IN ({array_int:messages})
-			AND content_type = {literal:post}',
-		array(
-			'messages' => $messages,
-		)
-	);
+	if (!empty($messages))
+	{
+		wesql::query('
+			DELETE FROM {db_prefix}likes
+			WHERE id_content IN ({array_int:messages})
+				AND content_type = {literal:post}',
+			array(
+				'messages' => $messages,
+			)
+		);
+	}
 
 	// Delete anything related to the topic.
 	wesql::query('
